Issue 27: Reset ERL_FLAGS before making rebar_utils:sh() calls
Make a wrapper for rebar_utils:sh() so that this plugin can invoke elixir/mix commands with ERL_FLAGS set to "". This is done so that the settings of ERL_FLAGS do not influence these commands.

For more information refer to: Issue 27, https://github.com/barrel-db/rebar3_elixir_compile/issues/27