Fix reading of mix lock file to not output warning
Elixir 1.7 started warning a lot when doing eval on keyword lists
that contained unnecessary quoting of the key. Unfortunately a lot
of mix lock files contain such atoms so a lot of warning were printed
to stdout. With this change the warnings are suppressed and a better
way to read the mix lock file is used.